+++ event-store-management/event-store-management-core/src/main/java/uk/gov/justice/services/eventstore/management/replay/commands/ReplayEventToEventListenerCommandHandler.java
@@ -2,7 +2,7 @@ package uk.gov.justice.services.eventstore.management.replay.commands;
 
 import org.slf4j.Logger;
 import uk.gov.justice.services.common.util.UtcClock;
-import uk.gov.justice.services.eventstore.management.replay.process.ReplayEventToEventListenerRunner;
+import uk.gov.justice.services.eventstore.management.replay.process.ReplayEventToComponentRunner;
 import uk.gov.justice.services.eventstore.management.commands.ReplayEventToEventListenerCommand;
 import uk.gov.justice.services.jmx.api.domain.CommandState;
 import uk.gov.justice.services.jmx.command.HandlesSystemCommand;
@@ -12,6 +12,7 @@ import javax.enterprise.event.Event;
 import javax.inject.Inject;
 import java.util.UUID;
 
+import static uk.gov.justice.services.core.annotation.Component.EVENT_LISTENER;
 import static uk.gov.justice.services.eventstore.management.commands.ReplayEventToEventListenerCommand.REPLAY_EVENT_TO_EVENT_LISTENER;
 import static uk.gov.justice.services.jmx.api.domain.CommandState.*;
 
@@ -21,7 +22,7 @@ public class ReplayEventToEventListenerCommandHandler {
     private Event<SystemCommandStateChangedEvent> stateChangedEventFirer;
 
     @Inject
-    private ReplayEventToEventListenerRunner replayEventToEventListenerRunner;
+    private ReplayEventToComponentRunner replayEventToComponentRunner;
 
     @Inject
     private UtcClock clock;
@@ -31,14 +32,14 @@ public class ReplayEventToEventListenerCommandHandler {
 
     @HandlesSystemCommand(REPLAY_EVENT_TO_EVENT_LISTENER)
     public void replayEventToEventListener(final ReplayEventToEventListenerCommand command, final UUID commandId, final UUID commandRuntimeId) {
-        fireEvent(COMMAND_IN_PROGRESS, command, commandId, "ReplayEventToEventListener command received");
+        fireEvent(COMMAND_IN_PROGRESS, command, commandId, "REPLAY_EVENT_TO_EVENT_LISTENER command received");
 
         try {
-            replayEventToEventListenerRunner.run(commandId, commandRuntimeId);
-            fireEvent(COMMAND_COMPLETE, command, commandId, "ReplayEventToEventListener command completed");
+            replayEventToComponentRunner.run(commandId, commandRuntimeId, EVENT_LISTENER);
+            fireEvent(COMMAND_COMPLETE, command, commandId, "REPLAY_EVENT_TO_EVENT_LISTENER command completed");
         } catch (Exception e) {
-            logger.error("ReplayEventToEventListenerCommand failed. commandId {}, commandRuntimeId {}", commandId, commandRuntimeId, e);
-            fireEvent(COMMAND_FAILED, command, commandId, "ReplayEventToEventListener command failed");
+            logger.error("REPLAY_EVENT_TO_EVENT_LISTENER failed. commandId {}, commandRuntimeId {}", commandId, commandRuntimeId, e);
+            fireEvent(COMMAND_FAILED, command, commandId, "REPLAY_EVENT_TO_EVENT_LISTENER command failed");
         }
     }